Update Refitter version in Visual studio context menus to v0.7.2
The button texts in VSCommandTable.vsct and Manifest.addin.xml files were updated with the new Refitter version (v0.7.2). This is a necessary update as the Refitter tool has been upgraded to a newer version and the corresponding changes need to be reflected in the Visual Studio and VS Mac extensions' UI buttons that trigger the use of this tool.