Update GitHub actions to use setup-java@v2
All instances of actions/setup-java@v1 have been replaced with actions/setup-java@v2 in the GitHub actions workflows. The change has been made to regression-tests.yml, sonar-cloud.yml, smoke-tests.yml, and production-tests.yml. This is to leverage the new features and bug fixes in version 2 and to keep our workflows up to date.