Update NSwag version in Visual Studio context menus to v13.20.0
Updated the NSwag version from v13.19.0 to v13.20.0 in the button text across `VSCommandTable.vsct` files & `Manifest.addin.xml` file. The change is necessary to reflect the correct version in the UI for consistency and to avoid any confusion for the users on the application version they are using for code generation.