Update OpenAPI Generator download link and hashes to use v7.0.0
Updated the OpenAPI Generator CLI download URL to point to the latest version (7.0.0) in the Resource.resx and Resource.Designer.cs files. Also, adjusted the associated MD5 and SHA1 hashes to match the new version's jar file. This is to ensure we are using the latest OpenAPI tools for our API client code generation.