Update OpenAPI Generator version to v7.0.0 in <auto-generated> code comment
The version of the OpenAPI Generator used in the project has been updated from 'v7.0.0-beta' to 'v7.0.0'. This is reflected in the code comment above the method 'MergeFilesAndDeleteSource'. This change aligns the comment with the actual version of the OpenAPI Generator being used. The amendment is necessary to maintain comment accuracy and the comprehensibility of the source code.