Update OpenAPI Generator version to 7.0.0 in Visual Studio context menus
The OpenAPI Generator version in command labels across multiple places in VSIX and VSMac folders are updated from 'v7.0.0-beta' to 'v7.0.0'. This reflects the transition from the beta version of OpenAPI Generator to the official release, helping the users with accurate version information in the UI of the tools.