Update JsonSubTypes and RestSharp package versions in integration tests
The package versions for JsonSubTypes and RestSharp in the OpenApiGeneratorProjectFileContents.cs file have been updated. This change ensures the application uses the latest versions of these packages. JsonSubTypes was updated from version 1.9.0 to 2.0.1, and RestSharp was updated from 108.0.2 to 110.2.0. It enhances functionality and security by including recent patches from the latest package versions.