Update JsonSubTypes package in mustache templates to version 2.0.1
The JsonSubTypes package was updated from version 1.9.0 to 2.0.1 across multiple project templates. This is to take advantage of the new features and bug fixes that have been released in the recent version of the package. It is crucial to maintain the latest versions for efficient project performance and to ensure compatibility with other packages.