Update RestSharp version across smoke test projects
The RestSharp package version has been updated from 108.0.2 to 110.2.0 in multiple csproj files. This change was necessary to incorporate the enhancements and bug fixes in the new version of RestSharp. Specifically, for the Net6 project, the Polly package version is also downgraded to 7.2.3 and Newtonsoft.Json package has been included.