Update JDK version in GitHub workflows
The JDK version used in the GitHub workflows for regression tests, Sonar Cloud analysis, smoke tests, and production tests has been updated from 11 to 17. This change was necessary to reflect the upgrade in the project's Java version to JDK 17, ensuring all tests and analyses run using the updated Java environment.