Update Refitter version in Visual Studio context menus to v0.7.3
The labels for the commands 'Generate with Refitter' in VSCommandTable.vsct and Manifest.addin.xml were updated to reflect the new version of Refitter (v0.7.3). This is a necessary update every time a new version of Refitter is released, to let the users know they are generating code with the latest version of the tool.