Refactor OpenApiCSharpCodeGenerator to support config files
The OpenApiCSharpCodeGenerator class in Core/ApiClientCodeGen.Core has been modified to support configuration files. This change is to enhance the flexibility of the code generation process. Now if a corresponding .config file exists for a given swagger file and the 'UseConfigurationFile' property is set to true, the generator will use the configuration file instead of the standard arguments. This opens up new configuration possibilities, such as defining custom additional properties per swagger file.