Update Refitter version in Visual Studio context menus to v0.7.5
This commit updates the version of Refitter indicated in the button labels within the VSIX and VSMac projects. The labels now accurately reflect that version 0.7.5 of Refitter is being used. This change ensures that users of the ApiClientCodeGen tool have the most up-to-date information available and do not try to generate code with an outdated version of Refitter.