Update Kiota version in VSIX and VSMac context menus
The button text indicating the version of Microsoft Kiota has been updated from v1.5.1 to v1.6.1. This change was made in the .vsct files of the VSIX project and the Manifest.addin.xml of the VSMac project. The update reflects the new version of Kiota being utilized and provides accurate information to the users using these tools.