Update OpenAPI generator URL and checksums
Updated the URL for the OpenAPI generator-download in the resource files, along with corresponding MD5 and SHA1 checksums. This needed adjustment due to upgrade OpenAPI generator from version 7.0.0 to 7.0.1, ensuring users are directed to the latest version on the Maven repository. This also demands changes in checksums to verify correct file integrity after the download.