Update OpenAPI Generator version in Visual Studio context menus
Through this commit, the versions mentioned in the buttonText for generating code with OpenAPI Generator in the ApiClientCodeGen toolsets have been updated from v7.0.0 to v7.0.1. The version update has been done to ensure users are aware of the latest version of the OpenAPI Generator being used for code generation. This change impacts both Visual Studio and VS Mac toolsets.