Update Refitter version in Visual Studio context menus
Updated the version of Refitter displayed in the button text in multiple areas throughout the application. The version was updated from v0.8.1 to v0.8.2. This was necessary to display accurate information to the users following the recent update of Refitter to version v0.8.2. Changes were made in VSIX and VSMac solutions, specifically in the ButtonText and _label attributes.