Enable logging for telemetry client in Debug/Test mode
This commit modifies the WriteLine method in AppInsightsRemoteLogger to perform logging when tests are running, or when the Debugger is attached, or if telemetry client is null. Previously, this method was intentionally left empty. This change will assist in troubleshooting and debugging by allowing monitoring of application insights data during these modes.