Fix Uri creation in GetStringAsync method
The way we were previously handling Uri creation in the GetStringAsync method was incorrect. In the EnterOpenApiSpecDialog.cs file, the line of code initializing the Uri with new Uri(tbUrl.Text) has been replaced by simply passing the needed tbUrl.Text as a string. This addresses the issue by passing the URL directly as a string without the need of wrapping it in a Uri object. This change is expected to resolve any potential exceptions related to URI parsing.