Update Kiota version in Visual Studio context menus
The button texts in 'VSCommandTable.vsct' for both ApiClientCodeGen.VSIX and ApiClientCodeGen.VSIX.Dev17, as well as 'Manifest.addin.xml' for ApiClientCodeGen.VSMac have been updated. The version of Microsoft Kiota has changed from v1.8.0 to v1.8.1. This change aligns the tool's version specified in the user interface with the latest version we're using for code generation.