Update Kiota version in documentation
The documentation has been updated to reflect the new version of the Kiota Code Generator, moving from v1.8.0 to v1.8.1. This change was necessary to ensure our documentation is up-to-date with the latest version of the Microsoft Kiota project, ensuring users accurately understand its current usage and functionality across our various tools and documentation sources.