Update Refitter version in Visual Studio context menus
The version of Refitter used in the button text across VSIX and VSMac files was updated from v0.8.3 to v0.8.4. It reflects the updated version of the Refitter library being used in the tool. It is essential for the button text to display the accurate version to inform users about the version of the library in use.