Update download link to Open API generator to use v7.1.0
The Open API generator version has been updated from 7.0.1 to 7.1.0 in both the download URL and the corresponding MD5 and SHA1 checksums in the Resource.resx file. This change is reflected in Resource.Designer.cs as well. This update ensures the code generator uses the recent version that may include bug fixes, improvements, and new features.