Update Polly NuGet package version in docs
The version of the Polly NuGet package listed in our documentation files (Marketplace2022.md, VisualStudioForMac.md, README.md, and Marketplace.md) was updated from 8.1.0 to 8.2.0. The change is tied to a recent upgrade in our dependencies to ensure that our users are referencing the most recent, stable version of the package for better performance and security.