Adjust dialog buttons and labels vertical positioning
Buttons and status labels within the EnterOpenApiSpecDialog were slightly off in their vertical positioning. This commit adjusts the Y-coordinate for the 'OK', 'Cancel', 'AddCustomHeaders' buttons and 'Status' , 'MarketplaceLink' labels  to ensure proper alignment and improve the visual appearance of the dialog. In addition, the overall ClientSize of the dialog has been increased to accommodate these changes without overlap.