Use the remote URL in .refitter created from New REST API Client -> Refitter context menu
The `NewRestClientCommand.cs` file has been significantly refactored. Namespaces have been reorganized with some removed and others added. Filename extensions are now converted to lowercase before use. The output namespace is retrieved earlier in the code. The filename is replaced with `.nswag` or `.refitter` depending on the `CodeGenerator` used.

A new `if` statement has been added to handle the `Refitter` CodeGenerator, creating a new `OptionsFactory` and `RefitGeneratorSettings` object. The logic for setting the `CustomTool` property of the `projectItem` has been moved to an `else` clause that only executes if the `CodeGenerator` is not `NSwagStudio` or `Refitter`.

An `else if` clause has been added to handle the `Refitter` CodeGenerator, creating a new `RefitterCodeGenerator` and calling its `GenerateCode` method. If the `CodeGenerator` is neither `NSwagStudio` nor `Refitter`, the `CustomTool` property of the `projectItem` is set to the custom tool name of the `CodeGenerator`.